Restore stack deploy integration test with dab
